Use Request.Host to get the host name
We want to report cluster address we are running in. The easier way to
do this without depending on Kubernetes API is to use the Host header in
the HTTP request, which is typically `http://cluster-host:node-port/`.

The previous method was reporting the pod network address (10.x.y.z)
which is not useful and always the same on minikube clusters.