Support optional cluster config
Try to read the cluster name from /mnt/config/cluster and use it instead
of the request Host header. This allow reporting the actual managed
cluster name instead of the IP address or the DNS name of the cluster.

Injecting the cluster name into a kubevirt vm can be done using a
configMap[1].

[1] https://kubevirt.io/user-guide/storage/disks_and_volumes/#as-a-disk_2