Fix deadlock when installing in cloud-final
If you start the service in cloud-config runcmd, the command never
complete since the service will start only after cloud-final starts, and
it starts only when cloud-config finish.

Our dependency is cloud-config when the config volume may be mounted.
With change we can enable and start the service in runcmd.